refactor(P2P): TECHDEBT:split structures in their respective file
