+++ p2p/types/transport.go
@@ -1,0 +1,8 @@
+package types
+
+type Transport interface {
+	IsListener() bool
+	Read() ([]byte, error)
+	Write([]byte) error
+	Close() error
+}
